requirements.txt
pillow == 10.*
